Generalize QVision to open space for other models and detection schemes
