+++ jansenlib/QVision/QVisionWidget.py
@@ -126,24 +126,24 @@ class Ui_QVisionWidget(object):
         self.checkFeatureData.setObjectName("checkFeatureData")
         self.formLayout_2.setWidget(3, QtWidgets.QFormLayout.LabelRole, self.checkFeatureData)
         self.verticalLayout.addWidget(self.groupExport)
-        self.plotZ = PlotWidget(QVisionWidget)
+        self.plot2 = PlotWidget(QVisionWidget)
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Preferred)
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(2)
-        sizePolicy.setHeightForWidth(self.plotZ.sizePolicy().hasHeightForWidth())
-        self.plotZ.setSizePolicy(sizePolicy)
-        self.plotZ.setObjectName("plotZ")
-        self.verticalLayout.addWidget(self.plotZ)
-        self.plotAN = PlotWidget(QVisionWidget)
-        self.plotAN.setEnabled(True)
+        sizePolicy.setHeightForWidth(self.plot2.sizePolicy().hasHeightForWidth())
+        self.plot2.setSizePolicy(sizePolicy)
+        self.plot2.setObjectName("plot2")
+        self.verticalLayout.addWidget(self.plot2)
+        self.plot1 = PlotWidget(QVisionWidget)
+        self.plot1.setEnabled(True)
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Preferred)
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(2)
-        sizePolicy.setHeightForWidth(self.plotAN.sizePolicy().hasHeightForWidth())
-        self.plotAN.setSizePolicy(sizePolicy)
-        self.plotAN.setStatusTip("")
-        self.plotAN.setObjectName("plotAN")
-        self.verticalLayout.addWidget(self.plotAN)
+        sizePolicy.setHeightForWidth(self.plot1.sizePolicy().hasHeightForWidth())
+        self.plot1.setSizePolicy(sizePolicy)
+        self.plot1.setStatusTip("")
+        self.plot1.setObjectName("plot1")
+        self.verticalLayout.addWidget(self.plot1)
 
         self.retranslateUi(QVisionWidget)
         QtCore.QMetaObject.connectSlotsByName(QVisionWidget)
