+++ jansenlib/QVision/__init__.py
@@ -1,0 +1,3 @@
+from .QHVM import QHVM as QVision
+
+all = [QVision]
